requirements.txt
asgiref==3.6.0
Django==4.1.5
django-rest-framework==0.1.0
django-restframework==0.0.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
mysql==0.0.3
mysqlclient==2.1.1
PyJWT==2.6.0
PyMySQL==1.0.2
pytz==2022.7.1
sqlparse==0.4.3
tzdata==2022.7
